Created the basis for a lot of the systems
Goes to the main menu and the player can choose a play. The audio for the title and space story is in and they both loop. The player can enable the text box with the T key and return to the main menu with the escape key when in one of the play rooms. The first line of dialog for one of the plays is on display. The basis for the dialog system is up as well.