Added a lot of sprites and finalized the functionality of the title and main menu screen
Main menu and title are working optimally but may need to add a button or two later down the line. Should not be too hard to implement. Added the sprites for the dialogue box and positioned it. Changed and responsed the text for the dialogue. Added all of the scenes and the rooms needed for the space story. Finally, found a way to determine when the dialogue ends and should go to the review room.