Fully functioning version of the game minus some sprites
The game is fully functional with the space story. The player can go in and get four different endings for the space story. Text is fully implemented and scrolls. The only addition needed are the sprites for the buttons. Finalized the ending scenes for the different endings and the speech bubbles.